[OpenShift] Refactors TektonAddon reconciler
This refactors addon reconciler and moves all components to individual
file rather than keeping everything at single place to improve
readability.
